+++ backend/canisters/local_user_index/api/src/queries/access_token_v2.rs
@@ -1,8 +1,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{BotActionScope, BotApiKeyToken, BotCommand, CanisterId, Chat, UserId, VideoCallType};
-use utils::base64;
+use types::{BotActionScope, BotCommand, Chat, UserId, VideoCallType};
 
 #[ts_export(local_user_index, access_token_v2)]
 #[allow(clippy::large_enum_variant)]
@@ -12,15 +11,7 @@ pub enum Args {
     JoinVideoCall(JoinVideoCallArgs),
     MarkVideoCallAsEnded(MarkVideoCallAsEndedArgs),
     BotActionByCommand(BotActionByCommandArgs),
-    BotActionByApiKey(#[serde(deserialize_with = "deserialize_bot_action_by_api_key_args")] BotApiKeyToken),
-}
-
-fn deserialize_bot_action_by_api_key_args<'de, D>(deserializer: D) -> Result<BotApiKeyToken, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let str = String::deserialize(deserializer)?;
-    base64::to_value(&str).map_err(serde::de::Error::custom)
+    BotActionByApiKey(String),
 }
 
 #[ts_export(local_user_index, access_token_v2)]
@@ -29,28 +20,6 @@ pub enum Response {
     Success(String),
     NotAuthorized,
     InternalError(String),
-}
-
-impl Args {
-    pub fn canister_id(&self) -> CanisterId {
-        match self {
-            Args::StartVideoCall(args) => args.chat.canister_id(),
-            Args::JoinVideoCall(args) => args.chat.canister_id(),
-            Args::MarkVideoCallAsEnded(args) => args.chat.canister_id(),
-            Args::BotActionByCommand(args) => args.scope.canister_id(),
-            Args::BotActionByApiKey(args) => args.scope.canister_id(),
-        }
-    }
-
-    pub fn type_name(&self) -> &str {
-        match self {
-            Args::StartVideoCall(_) => "StartVideoCall",
-            Args::JoinVideoCall(_) => "JoinVideoCall",
-            Args::MarkVideoCallAsEnded(_) => "MarkVideoCallAsEnded",
-            Args::BotActionByCommand(_) => "BotActionByCommand",
-            Args::BotActionByApiKey(_) => "BotActionByApiKey",
-        }
-    }
 }
 
 #[ts_export]