+++ backend/libraries/group_chat_core/src/lib.rs
@@ -620,7 +620,7 @@ impl GroupChatCore {
         }
 
         // If there is an existing message with the same message id then this is invalid unless
-        // a bot is finalising its unfinalised message
+        // a bot is updating an unfinalised message
         if let Some((message, _)) =
             self.events
                 .message_internal(EventIndex::default(), thread_root_message_index, message_id.into())
@@ -628,13 +628,13 @@ impl GroupChatCore {
             if let Caller::BotV2(bot_now) = &caller {
                 if let Some(bot_message) = message.bot_context {
                     if bot_now.bot == message.sender
-                        && bot_now.command.initiator == bot_message.command.initiator
+                        && bot_now.command.as_ref().map(|c| c.initiator) == bot_message.command.as_ref().map(|c| c.initiator)
                         && bot_now.command == bot_message.command
-                        && bot_now.finalised
                         && !bot_message.finalised
                     {
-                        return self.finalise_bot_message(
+                        return self.update_bot_message(
                             caller,
+                            bot_now.finalised,
                             thread_root_message_index,
                             message_id,
                             content,
@@ -749,9 +749,10 @@ impl GroupChatCore {
         })
     }
 
-    fn finalise_bot_message(
+    fn update_bot_message(
         &mut self,
         caller: &Caller,
+        finalise: bool,
         thread_root_message_index: Option<MessageIndex>,
         message_id: MessageId,
         content: MessageContentInitial,
@@ -789,7 +790,7 @@ impl GroupChatCore {
             message_id,
             content,
             block_level_markdown: Some(block_level_markdown),
-            finalise_bot_message: true,
+            finalise_bot_message: finalise,
             now,
         };
 
@@ -802,21 +803,25 @@ impl GroupChatCore {
 
         let message_event = reader.message_event(message_id.into(), Some(caller.agent())).unwrap();
 
-        let users_to_notify = self.build_users_to_notify(
-            thread_root_message_index,
-            min_visible_event_index,
-            replies_to,
-            &message_event,
-            mentioned,
-            suppressed,
-            everyone_mentioned,
-            now,
-        );
+        let users_to_notify = if finalise {
+            self.build_users_to_notify(
+                thread_root_message_index,
+                min_visible_event_index,
+                replies_to,
+                &message_event,
+                mentioned,
+                suppressed,
+                everyone_mentioned,
+                now,
+            )
+        } else {
+            vec![]
+        };
 
         SendMessageResult::Success(SendMessageSuccess {
             message_event,
             users_to_notify,
-            unfinalised_bot_message: false,
+            unfinalised_bot_message: !finalise,
         })
     }
 
@@ -936,33 +941,42 @@ impl GroupChatCore {
             });
         }
 
-        let member = match self.members.get_verified_member(caller.initiator()) {
-            Ok(member) => member,
-            Err(error) => {
-                return match error {
-                    VerifyMemberError::NotFound => UserNotInGroup,
-                    VerifyMemberError::Lapsed => UserLapsed,
-                    VerifyMemberError::Suspended => UserSuspended,
-                };
+        let (min_visible_event_index, can_mention) = if let Some(initator) = caller.initiator() {
+            let member = match self.members.get_verified_member(initator) {
+                Ok(member) => member,
+                Err(error) => {
+                    return match error {
+                        VerifyMemberError::NotFound => UserNotInGroup,
+                        VerifyMemberError::Lapsed => UserLapsed,
+                        VerifyMemberError::Suspended => UserSuspended,
+                    };
+                }
+            };
+
+            if !matches!(content, MessageContentInternal::VideoCall(_)) && !member.check_rules(&self.rules.value) {
+                return RulesNotAccepted;
             }
+
+            let permissions = &self.permissions;
+
+            if !member
+                .role()
+                .can_send_message(content.into(), thread_root_message_index.is_some(), permissions)
+            {
+                return NotAuthorized;
+            }
+
+            (
+                member.min_visible_event_index(),
+                member.role().can_mention_everyone(permissions),
+            )
+        } else {
+            (EventIndex::default(), true)
         };
 
-        if !matches!(content, MessageContentInternal::VideoCall(_)) && !member.check_rules(&self.rules.value) {
-            return RulesNotAccepted;
-        }
-
-        let permissions = &self.permissions;
-
-        if !member
-            .role()
-            .can_send_message(content.into(), thread_root_message_index.is_some(), permissions)
-        {
-            return NotAuthorized;
-        }
-
         Success(PrepareSendMessageSuccess {
-            min_visible_event_index: member.min_visible_event_index(),
-            everyone_mentioned: member.role().can_mention_everyone(permissions) && is_everyone_mentioned(content),
+            min_visible_event_index,
+            everyone_mentioned: can_mention && is_everyone_mentioned(content),
         })
     }